requirements.txt
numpy==1.24.3
opencv-python==4.8.0.74
scikit-image==0.21.0
matplotlib==3.7.2
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
scipy==1.11.2
pandas==2.0.3
